requirements.txt
django
psycopg2
djangorestframework
django-environ
dj-database-url
requests-oauthlib==1.3.1
